// packages/shepherd-backend/src/playlists/types/playlist.domain.types.ts
import type {ObjectId} from "mongodb";

/**
 * Interface representing a complete playlist, generated by `PlaylistBuilder`.
 */
export type TPlaylistDomainModel = {
    /** General playlist settings */
    settings: TPlaylistSettings;
    tags: string[];
    /** List of songs in the playlist */
    songList: TPlaylistSong[];

    /** Configuration of the performers */
    performers: {
        singersConfig: TSingersConfig;
        musiciansConfig: TMusicianConfig;
        aerialConfig: TAerialConfig;
    };
    dataInformation: TDataInformation;
}


export type TPlaylistSettings = {
    name: string;
    description: string;
    usage: 'daily' | 'event';
    energy: 1 | 2 | 3 | 4;
    requiredLength: number;
    numberOfSongs: number;
    singers: boolean;
    musicians: boolean;
    aerial: boolean;
}

export type TSingersConfig = {
    numberOfSingers: number | null;
    containsDuo: boolean | null;
    splitMode: "alternate" | "equal_split" | null;
}

export type TAerialConfig = { performancePosition: "start" | "end" | "manual" | null };
export type TMusicianConfig = { role: "solo" | "duo" | null };

export type TPlaylistSong = {
    _id: string | ObjectId | null;
    title: string | null;
    artist: string | null;
    version_id: string | ObjectId | null;
    version_length: number | null;
    tags: string[];
}

export type TDataInformation = {
    creation_date: Date;
    creator_id: ObjectId;
    last_modified: Date;
    updateNumber: number;
}